requirements.txt
aniso8601==9.0.1
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-RESTful==0.3.9
flask-restx==1.0.3
flask-swagger==0.2.14
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.0
MarkupSafe==2.1.1
Pillow==9.3.0
pyrsistent==0.19.2
pyTelegramBotAPI==4.7.1
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
